requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-environ==0.12.0
django-js-asset==3.1.2
django-phonenumber-field==8.1.0
django-query-counter==0.4.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
phonenumbers==9.0.10
pillow==11.3.0
psycopg==3.2.9
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.14.1
uritemplate==4.2.0
